requirements.txt
requests
pandas
python-dotenv
sqlalchemy
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
